LayoutTests/imported/w3c/web-platform-tests/workers/importscripts_mime_local.any.sharedworker-expected.txt


PASS importScripts() requires scripty MIME types for data: URLs: text/javascript is allowed.
PASS importScripts() requires scripty MIME types for data: URLs: application/javascript is allowed.
PASS importScripts() requires scripty MIME types for data: URLs: text/ecmascript is allowed.
PASS importScripts() requires scripty MIME types for data: URLs: image/png is blocked.
PASS importScripts() requires scripty MIME types for data: URLs: text/csv is blocked.
PASS importScripts() requires scripty MIME types for data: URLs: video/mpeg is blocked.
PASS importScripts() requires scripty MIME types for data: URLs: text/html is blocked.
PASS importScripts() requires scripty MIME types for data: URLs: text/plain is blocked.
PASS importScripts() requires scripty MIME types for data: URLs: application/xml is blocked.
PASS importScripts() requires scripty MIME types for data: URLs: application/octet-stream is blocked.
PASS importScripts() requires scripty MIME types for data: URLs: text/potato is blocked.
PASS importScripts() requires scripty MIME types for data: URLs: potato/text is blocked.
PASS importScripts() requires scripty MIME types for data: URLs: aaa/aaa is blocked.
PASS importScripts() requires scripty MIME types for data: URLs: zzz/zzz is blocked.
PASS importScripts() requires scripty MIME types for data: URLs: text/javascript; charset=utf-8 is allowed.
PASS importScripts() requires scripty MIME types for data: URLs: text/javascript;charset=utf-8 is allowed.
PASS importScripts() requires scripty MIME types for data: URLs: text/javascript;bla;bla is allowed.
PASS importScripts() requires scripty MIME types for data: URLs: text/csv; charset=utf-8 is blocked.
PASS importScripts() requires scripty MIME types for data: URLs: text/csv;charset=utf-8 is blocked.
PASS importScripts() requires scripty MIME types for data: URLs: text/csv;bla;bla is blocked.
PASS importScripts() requires scripty MIME types for data: URLs: Text/html is blocked.
PASS importScripts() requires scripty MIME types for data: URLs: text/Html is blocked.
PASS importScripts() requires scripty MIME types for data: URLs: TeXt/HtMl is blocked.
PASS importScripts() requires scripty MIME types for data: URLs: TEXT/HTML is blocked.
PASS importScripts() requires scripty MIME types for blob: URLs: text/javascript is allowed.
PASS importScripts() requires scripty MIME types for blob: URLs: application/javascript is allowed.
PASS importScripts() requires scripty MIME types for blob: URLs: text/ecmascript is allowed.
PASS importScripts() requires scripty MIME types for blob: URLs: image/png is blocked.
PASS importScripts() requires scripty MIME types for blob: URLs: text/csv is blocked.
PASS importScripts() requires scripty MIME types for blob: URLs: video/mpeg is blocked.
PASS importScripts() requires scripty MIME types for blob: URLs: text/html is blocked.
PASS importScripts() requires scripty MIME types for blob: URLs: text/plain is blocked.
PASS importScripts() requires scripty MIME types for blob: URLs: application/xml is blocked.
PASS importScripts() requires scripty MIME types for blob: URLs: application/octet-stream is blocked.
PASS importScripts() requires scripty MIME types for blob: URLs: text/potato is blocked.
PASS importScripts() requires scripty MIME types for blob: URLs: potato/text is blocked.
PASS importScripts() requires scripty MIME types for blob: URLs: aaa/aaa is blocked.
PASS importScripts() requires scripty MIME types for blob: URLs: zzz/zzz is blocked.
PASS importScripts() requires scripty MIME types for blob: URLs: text/javascript; charset=utf-8 is allowed.
PASS importScripts() requires scripty MIME types for blob: URLs: text/javascript;charset=utf-8 is allowed.
PASS importScripts() requires scripty MIME types for blob: URLs: text/javascript;bla;bla is allowed.
PASS importScripts() requires scripty MIME types for blob: URLs: text/csv; charset=utf-8 is blocked.
PASS importScripts() requires scripty MIME types for blob: URLs: text/csv;charset=utf-8 is blocked.
PASS importScripts() requires scripty MIME types for blob: URLs: text/csv;bla;bla is blocked.
PASS importScripts() requires scripty MIME types for blob: URLs: Text/html is blocked.
PASS importScripts() requires scripty MIME types for blob: URLs: text/Html is blocked.
PASS importScripts() requires scripty MIME types for blob: URLs: TeXt/HtMl is blocked.
PASS importScripts() requires scripty MIME types for blob: URLs: TEXT/HTML is blocked.
